+++ lib/oli/delivery/activity_provider.ex
@@ -157,6 +157,12 @@ defmodule Oli.Delivery.ActivityProvider do
 
           {c_errors ++ errors, c_activities ++ activities, [e | model], source}
 
+        "survey" ->
+          {c_errors, c_activities, c_model, source} = fulfill(e["children"], source)
+          e = %{e | "children" => Enum.reverse(c_model)}
+
+          {c_errors ++ errors, c_activities ++ activities, [e | model], source}
+
         _ ->
           {errors, activities, [e | model], source}
       end